my fixes

+++ utils/training_cocoop.py
@@ -59,7 +59,6 @@ def training_step(model, dataset, optimizer, batch_size, device="cuda"):
     tmp_dataset = ContiguousLabelDataset(dataset)
     dataloader = DataLoader(tmp_dataset, batch_size=batch_size, shuffle=True, num_workers=2)
     pbar = tqdm(dataloader, desc="Training", position=1, leave=False)
-
     for batch_idx, (inputs, targets) in enumerate(dataloader):
         # Load data into GPU
         inputs = inputs.to(device)
@@ -68,10 +67,13 @@ def training_step(model, dataset, optimizer, batch_size, device="cuda"):
         # Forward pass + loss computation
         logits, loss = model(inputs, targets)
 
+        # Backward pass
         loss.backward()
 
+        # Parameters update
         optimizer.step()
 
+        # Gradients reset
         optimizer.zero_grad()
 
         # Fetch prediction and loss value
@@ -137,7 +139,7 @@ def base_test_step(model: CLIP, dataset, categories, batch_size, device, label="
 
     # we can encode the text features once as they are shared for all images
     # therefore we do it outside the evaluation loop
-    text_features = model.encode_text(text_inputs).float()
+    text_features = model.encode_text(text_inputs)
     # and here we normalize them (standard pratice with CLIP)
     text_features /= text_features.norm(dim=-1, keepdim=True)
 
@@ -157,7 +159,7 @@ def base_test_step(model: CLIP, dataset, categories, batch_size, device, label="
         target = target.to(device)
 
         # forward image through CLIP image encoder
-        image_features = model.encode_image(image.type(torch.float32))
+        image_features = model.encode_image(image)
         # and normalize
         image_features /= image_features.norm(dim=-1, keepdim=True)
